+++ CHANGELOG.md
@@ -1,7 +1,38 @@
+## [0.2.1] (2025/01/19) - Remembrance
+
+### Changes
+- Overall changes to program execution. A few notable highlights:
+  - Execution history is no longer stored as a list. A print statement must be
+    used in its place for logging.
+  - Program execution now occurs in the IO monad, as more operations involving
+    I/O will be implemented in the future.
+  - Slight refactor and renames to highlight the fetch-decode-execute cycle.
+  - To run programs, one must invoke `runProgram` instead of `execute`
+- Changed the behavior of `JEQ`, `JNQ`, `JGT` and `JLT` to compare with the
+  newly added register.
+- Renamed `CPU` to `ExecUnit`. 
+- Replaced the `state` field in `ExecUnit` with the `instr` field, which
+  contains the currently executed instruction.
+- Renamed `emptyCPU` to `initState`.
+
+### Removals
+- Removed `JQV`, `JNV`, `JGV` and `JLV`. With the addition of the register,
+  these opcodes were unnecessary.
+- Removed the dependency for `Data.Text` in `Cascade.Runtime`. This may be
+  reverted in the future.
+
+### Additions
+- Added the `reg` field in `ExecUnit`, which represents a program register.
+- Added `LVR`, `RGS`, `SRG` and `JMR`, which involves the register.
+- Added `LMR` and `SMR`, though these two instructions currently do nothing.
+- With a complete refactoring of program execution, added more sensible runtime
+  error handling with `RuntimeError` and `Status`, indicating the current
+  status of the program.
+
 ## [0.1.2] (2024/11/22) - Twofold
 
 ### Additions
-- Added bitwise instructions `AND`, `IOR`, `XOR`, `NOT`, `SHL`, `SHR`, and `CNT`
+- Added bitwise instructions `AND`, `IOR`, `XOR`, `NOT`, `SHL`, `SHR` and `CNT`
 
 ## [0.1.1] (2024/11/02) - Divergence
 
@@ -18,7 +49,7 @@
 - Cleared `Cascade.Assembler`; a new assembler will be implemented soon
 
 ### Additions
-- Added various branching instructions, as well as `DUP`, `MUL`, `DIV`, and
+- Added various branching instructions, as well as `DUP`, `MUL`, `DIV` and
   `MOD`; for more information, check `README.md`
 - Added the instruction pointer to `CPU`